+++ src/libddot.f90
@@ -30,7 +30,7 @@ contains
 
       type(Tensor2), intent(in) :: T1
       type(Tensor2), intent(in) :: T2
-      real(kind=8) :: ddot_22
+      real(kind=dp) :: ddot_22
       integer :: i, j
 
       ddot_22 = 0.d0
@@ -47,7 +47,7 @@ contains
 
       type(Tensor2s), intent(in) :: T1
       type(Tensor2s), intent(in) :: T2
-      real(kind=8) :: ddot_2s2s
+      real(kind=dp) :: ddot_2s2s
       integer :: i
 
       ddot_2s2s = 0.d0
@@ -88,7 +88,7 @@ contains
       type(Tensor2s), intent(in) :: T1
       type(Tensor4s), intent(in) :: T2
       type(Tensor2s) :: ddot_2s4s
-      real(kind=8) :: w
+      real(kind=dp) :: w
       integer :: i, j
 
       ddot_2s4s%a6 = 0.d0
@@ -134,7 +134,7 @@ contains
       type(Tensor4s), intent(in) :: T1
       type(Tensor2s), intent(in) :: T2
       type(Tensor2s) :: ddot_4s2s
-      real(kind=8) :: w
+      real(kind=dp) :: w
       integer :: i, j
 
       ddot_4s2s%a6 = 0.d0
@@ -183,7 +183,7 @@ contains
       type(Tensor4s), intent(in) :: T1
       type(Tensor4s), intent(in) :: T2
       type(Tensor4s) :: ddot_4s4s
-      real(kind=8) :: w
+      real(kind=dp) :: w
       integer :: i, j, k
 
       ddot_4s4s%a6b6 = 0.d0
