+++ src/libdet.f90
@@ -9,7 +9,7 @@ contains
       implicit none
 
       type(Tensor2) :: T
-      real(kind=8) :: det_2
+      real(kind=dp) :: det_2
 
       det_2 = T%ab(1, 1)*(T%ab(2, 2)*T%ab(3, 3) - T%ab(2, 3)*T%ab(3, 2)) &
               + T%ab(1, 2)*(T%ab(2, 3)*T%ab(3, 1) - T%ab(2, 1)*T%ab(3, 3)) &
@@ -21,7 +21,7 @@ contains
       implicit none
 
       type(Tensor2s), intent(in) :: T
-      real(kind=8) :: det_2s
+      real(kind=dp) :: det_2s
 
       det_2s = T%a6(1)*T%a6(2)*T%a6(3) &
                + T%a6(4)*T%a6(5)*T%a6(6) &
