+++ src/libreducedim.f90
@@ -10,7 +10,7 @@ contains
 
       type(Tensor2s), intent(in) :: T
       integer, intent(in) :: i
-      real(kind=8), dimension(i) :: reduce_dim_2s
+      real(kind=dp), dimension(i) :: reduce_dim_2s
 
       reduce_dim_2s = T%a6(1:i)
 
@@ -21,7 +21,7 @@ contains
 
       type(Tensor4s), intent(in) :: T
       integer, intent(in) :: i, j
-      real(kind=8), dimension(i, j) :: reduce_dim_4s
+      real(kind=dp), dimension(i, j) :: reduce_dim_4s
 
       reduce_dim_4s = T%a6b6(1:i, 1:j)
