+++ src/libunimodular.f90
@@ -16,7 +16,7 @@ contains
 
       type(Tensor2), intent(in) :: T
       type(Tensor2) :: unimod_2, Eye
-      real(kind=8) :: detT
+      real(kind=dp) :: detT
 
       detT = det(T)
       Eye = identity2(Eye)
@@ -30,7 +30,7 @@ contains
 
       type(Tensor2s), intent(in) :: T
       type(Tensor2s) :: unimod_2s, Eye
-      real(kind=8) :: detT
+      real(kind=dp) :: detT
 
       detT = det(T)
       Eye = identity2(Eye)
@@ -44,7 +44,7 @@ contains
 
       type(Tensor2), intent(in) :: T
       type(Tensor2) :: unimod_2d, Eye
-      real(kind=8), intent(in) :: detT
+      real(kind=dp), intent(in) :: detT
 
       Eye = identity2(Eye)
 
@@ -57,7 +57,7 @@ contains
 
       type(Tensor2s), intent(in) :: T
       type(Tensor2s) :: unimod_2sd, Eye
-      real(kind=8), intent(in) :: detT
+      real(kind=dp), intent(in) :: detT
 
       Eye = identity2(Eye)
